backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
apscheduler
python-dotenv
sendgrid
twilio
